requirements.txt
ftfy
regex
tqdm
numpy
Pillow
scipy

termcolor==1.1.0
open3d==0.14.1
pgpelib @ git+https://github.com/nnaisense/pgpelib.git@d00907fe4b237e875ac0cc5b9d9c22591e4a7206
clip @ git+https://github.com/openai/CLIP.git@3b473b0e682c091a9e53623eebc1ca1657385717